PD#87250:clean warning for kernel
commit 74c66196b4aadad3eb9fa8baacbd137cd3e448e1
Author: zhi.zhou <zhi.zhou@amlogic.com>
Date:   Fri Feb 21 18:52:58 2014 +0800

    PD#87250 amports:clean up compile warning.

    Conflicts:

    	drivers/amlogic/amports/video.c

commit d0ea55670567434c8ad605d6f43c49f5ef67a7fe
Author: chengshun.wang <chengshun.wang@amlogic.com>
Date:   Thu Feb 20 19:21:26 2014 +0800

    PD# 87250: [audio codec]: clean up compile warning

commit b96b241eb54dbe467e29fcf291c67ffb8df4ddcb
Author: zedong.xiong <zedong.xiong@amlogic.com>
Date:   Thu Feb 20 16:56:47 2014 +0800

    PD 87250:amaudio and audiodsp:clean up compile warning.

commit 1858fa064392d5d26642e90bb46afbdac7fc7041
Author: tao.zeng <tao.zeng@amlogic.com>
Date:   Thu Feb 20 09:56:09 2014 +0800

    PD #87250: Remove compile warning for PMU driver.

commit 8dc84c669195437b247a60cab62f83d48fb8c8c9
Author: shi.liu <shi.liu@amlogic.com>
Date:   Thu Feb 20 10:51:17 2014 +0800

    PD:#87250:cpufreq:  remove warning

commit 5dbc77c960c2f9b1ddf011272ff1d9e6b785201a
Author: xiaoyu.yuan <xiaoyu.yuan@droid05.amlogic.com>
Date:   Thu Feb 20 10:18:15 2014 +0800

    PD #87250 : remove config parse warning message

commit dc197355d9d24da2a2495f7595a3b30bacb30e07
Author: benlong.zhou <benlong.zhou@amlogic.com>
Date:   Wed Feb 19 20:12:04 2014 +0800

    efuse: modified CLK_GATE_ON warning

commit c57f4e7d19ccf24c1be584d1a493666b8eca6cb2
Author: wei.nie <wei.nie@amlogic.com>
Date:   Wed Feb 19 19:31:28 2014 +0800

    PD #87250:OSD&LOGO clean up compile warning

commit 5b423fb0267cd2070ef6e2db0611251a020efa94
Author: zhenfei.li <zhenfei.li@amlogic>
Date:   Wed Feb 19 18:59:14 2014 +0800

    PD #87250: rm warings of aml_rtc driver.

commit 516b0859499926e637ce362ae7d1ff7d71108397
Author: zhenfei.li <zhenfei.li@amlogic>
Date:   Wed Feb 19 18:42:07 2014 +0800

    PD #87250: remove the warnings of the cameras.


Former-commit-id: e146c275373a0d3ec0433e36912de8c895d21aac

+++ drivers/amlogic/amports/subtitle.c
@@ -437,12 +437,12 @@ static struct class_attribute subtitle_class_attrs[] = {
 	__ATTR(reset, 	S_IRUGO | S_IWUSR, show_reset,  store_reset),
     __ATTR_NULL
 };
-
+/*
 static struct class subtitle_class = {
         .name = "subtitle",
         .class_attrs = subtitle_class_attrs,
     };
-
+*/
 /*********************************************************
  * /dev/amvideo APIs
  *********************************************************/
@@ -526,7 +526,7 @@ static long amsubtitle_ioctl(struct file *file,
                 break;
             case SUB_DATA_T_DATA: {
                     if (states->subtitle_info > 0) {
-						states->subtitle_info = subtitle_data[subtitle_write_pos].data;
+                        states->subtitle_info = (int)subtitle_data[subtitle_write_pos].data;
                     }
 				}
                 break;
